feat: adicionando filtro de pessoa na empresa e corrigindo associacao de pessoa com empresa

+++ app/Services/Person/PersonUpdateService.php
@@ -4,27 +4,37 @@ namespace App\Services\Person;
 
 use App\Models\Person;
 use App\Services\StorageDocumentService;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
 class PersonUpdateService
 {
     public function __construct(
         protected StorageDocumentService $storageDocumentService
-    ) {}
+    )
+    {
+    }
 
     public function update($id, array $data)
     {
         $person = Person::findOrFail($id);
-        if (isset($data['password'])) {
-            $data['password'] = bcrypt($data['password']);
-        }
-        if (isset($data['document'])) {
-            /// Deleta o arquivo antigo
-            $data = $this->storageDocumentService->storageDocument($data);
-            Storage::disk('s3')->delete($person->document_path);
-        }
-        $person->update($data);
-        return $person;
+        return DB::transaction(function () use ($person, $data) {
+            if (isset($data['password'])) {
+                $data['password'] = bcrypt($data['password']);
+            }
+            if (isset($data['document'])) {
+                /// Deleta o arquivo antigo
+                $data = $this->storageDocumentService->storageDocument($data);
+                if($person->document_path) {
+                    Storage::disk('s3')->delete($person->document_path);
+                }
+            }
+            $person->update($data);
+            if(isset($data['companies_id'])) {
+                $person->companies()->sync($data['companies_id']);
+            }
+            return $person;
+        });
     }
 }
